Update events (#42)
- Add March monthly meet

- Refactor data/template to remove duplication and avoid repeating the
same info everywhere.
    - If no location is specified, default to "Denver, CO"
- If no name is specified, auto-generate "Suspension Sunday" or
"Suspension Saturday" based on the `start_date`

- Sort events by date in Liquid

- Display location below the date (I think this looks better and makes
more sense).

- Remove filtering for old events. This didn't seem worth the trouble
because it is not dynamic. The page will only update when we push
changes -- i.e., after we edit `events.yml`. Since we have to do that,
we might as well manually delete old events when doing that.

## Screenshot

![Screenshot 2025-01-13 at 7 22
49 PM](https://github.com/user-attachments/assets/b3ca8c91-53e6-47f6-8aaa-5370bb77b00e)