Add anchors for headings
This adds anchor tags for all headings that use `heading_block.html`.

This allows for jumping/linking to specific sections on a page.

For example, the url `denversuspension.com/retreats/2025/#registration-pricing` will load the page and jump directly to the "Registration & Pricing" section.

This is a "nice to have" that I think will be particularly use for all the Resources pages.